Removed unused ObjectWriter and ObjectReader in AbstractIntegrationService.java.
Instead of having separate instances of ObjectWriter and ObjectReader, the ObjectMapper instance is now directly used for reading and writing operations.

Modified ActionType.java and Type.java to have @Getter annotation at class level for simplicity.

Enhanced the test case for subscriptions in SubscriptionServiceTest.java by adding an assertion to verify the non-nullity of the subscription list returned.

Updated SubscriptionService.java to filter the list of subscriptions returned based on the subscription type specified.

These changes aim to improve the management and usage of subscriptions from FIWARE in the application, by enabling accurate reading of subscription particulars and asserting correct behavior.