Add subscription removal feature and tests in SubscriptionService.
Added the ability to delete subscriptions from the FIWARE server in SubscriptionService.java. This is for the better management of the subscriptions within the application. Also added corresponding tests in SubscriptionServiceTest.java to ensure that the addition works as expected. These tests create and remove mock subscriptions and check that the number of existing subscriptions match the expected values after the operations. Improvement of the test coverage for the subscription functionalities is also intended with these additions.