Feature/enable and read subscriptions (#15)
* Enable and read subscriptions from FIWARE.

* Removed unused ObjectWriter and ObjectReader in AbstractIntegrationService.java.

Instead of having separate instances of ObjectWriter and ObjectReader, the ObjectMapper instance is now directly used for reading and writing operations.

Modified ActionType.java and Type.java to have @Getter annotation at class level for simplicity.

Enhanced the test case for subscriptions in SubscriptionServiceTest.java by adding an assertion to verify the non-nullity of the subscription list returned.

Updated SubscriptionService.java to filter the list of subscriptions returned based on the subscription type specified.

These changes aim to improve the management and usage of subscriptions from FIWARE in the application, by enabling accurate reading of subscription particulars and asserting correct behavior.

* Add subscription removal feature and tests in SubscriptionService.

Added the ability to delete subscriptions from the FIWARE server in SubscriptionService.java. This is for the better management of the subscriptions within the application. Also added corresponding tests in SubscriptionServiceTest.java to ensure that the addition works as expected. These tests create and remove mock subscriptions and check that the number of existing subscriptions match the expected values after the operations. Improvement of the test coverage for the subscription functionalities is also intended with these additions.

* Add subscription removal feature to improve management

SubscriptionService.java now has the ability to delete subscriptions from the FIWARE server, a change made to enhance how subscriptions are managed within the app. The corresponding tests added in SubscriptionServiceTest.java verify the correct workings of this function by creating and subsequently removing mock subscriptions. These tests confirm the count of active subscriptions against the expected figures. Enhancing test coverage for subscription functionalities is an expected benefit of these changes.

* Rename SubscriptionServiceTest to SubscriptionServiceIT

The class SubscriptionServiceTest.java has been renamed to SubscriptionServiceIT.java. This change aims to reflect the naming standard where IT stands for Integration Test. This feature confirms a clear distinction between unit tests and integration tests, improving maintenance and understanding of the test hierarchy.