Update DeviceIntegrationService URLs and add new Service classes (#17)
* Update DeviceIntegrationService URLs and add new Service classes

Changed the URLs used in DeviceIntegrationService to properly route requests to "http://localhost:1026" instead of "http://localhost:1026/v2". Also a StatusService class was added which retrieves version information from the context broker. Similarly, a Version model was added to map the version information.

AbstractIntegrationService now has functions for returning the context broker's base URL and command URL separately. Integration tests were also updated accordingly. These changes improve the modularity and readability of the code, making the service classes more dynamic and reusable.

* Update pom.xml version to 4.5.0

The version number of the 'fiware-integration-layer' project in pom.xml file is updated from 4.4.0 to 4.5.0. The update is made in response to recent enhancements to the project including URL changes in Device Integration Service and addition of new Service classes.