Update version in pom.xml file
Bumped up the version from 5.4.0 to 6.0.0 in pom.xml. This Version change reflects the recent updates made to the project, including the implementation of tenant support in Fiware integration services for resource organization and isolation. As part of these updates, also made necessary changes to the corresponding test resources and services.