Add CustomHeader for tenant identification in Fiware API calls
Introduced a 'CustomHeader' interface which standardizes the use of the 'fiware-service' HTTP header across the application. It improves code consistency, reduces potential human error and enhances code readability. This header carries the identity of the tenant(service) and eases resource coordination in multi-tenancy environments. Changes are made in the AbstractEntityIntegrationService and SubscriptionService classes, where the header is being added to the HTTP requests.