Feature/add multitenancy layer for fiware calls (#22)
* Added tenant support to Fiware integration services

Implemented tenant support in Fiware integration services to organize resources and provide resource isolation. Integrated this implementation into the existing test suites. Made necessary amendments to the corresponding test resources and instantiated services with the tenant parameter. Removed StatusServiceIT test file as the functionality is covered in DeviceIntegrationServiceBasedOnTenantIT.

* Update version in pom.xml file

Bumped up the version from 5.4.0 to 6.0.0 in pom.xml. This Version change reflects the recent updates made to the project, including the implementation of tenant support in Fiware integration services for resource organization and isolation. As part of these updates, also made necessary changes to the corresponding test resources and services.

* Add CustomHeader for tenant identification in Fiware API calls

Introduced a 'CustomHeader' interface which standardizes the use of the 'fiware-service' HTTP header across the application. It improves code consistency, reduces potential human error and enhances code readability. This header carries the identity of the tenant(service) and eases resource coordination in multi-tenancy environments. Changes are made in the AbstractEntityIntegrationService and SubscriptionService classes, where the header is being added to the HTTP requests.