Added functionality to delete a device from the system
Imported FiwareIntegrationLayerException and added a delete function to the AbstractEntityIntegrationService class to allow deletion of a device from the system by id. A test has been written in DeviceIntegrationServiceIT to ensure its proper operation. This change was made to improve the device management, where we can now delete devices that are no longer used or invalid.