Feature/add location to device for direct data logging (#25)
* Add location attribute in Device model

Added a location attribute to the Device model to provide more detailed information about devices. The 'location' attribute will allow to track the whereabouts of a device, enhancing the application's device management capabilities. The DeviceCategory class was minimally modified for optimal code formatting.

* Update version in pom.xml

The version of the fiware-integration-layer in the pom.xml file has been updated from 6.0.0 to 6.1.0. This change reflects the addition of the 'location' attribute to the Device model, which enhances device tracking and management capabilities. Minor modifications were also made to the DeviceCategory class for improved code formatting.

* Added functionality to delete a device from the system

Imported FiwareIntegrationLayerException and added a delete function to the AbstractEntityIntegrationService class to allow deletion of a device from the system by id. A test has been written in DeviceIntegrationServiceIT to ensure its proper operation. This change was made to improve the device management, where we can now delete devices that are no longer used or invalid.