Add manufacturerSpecificId to Device model and tests (#29)
      * Add manufacturerSpecificId to Device model and tests

Updated the Device model to include the 'manufacturerSpecificId' field and modified all the 'DeviceIntegrationService*' and 'DroneDeviceMeasurementIntegrationService' integration tests to include the 'manufacturerSpecificId' in their respective Device builder. This change ensures more detailed device identification and improved traceability in device-related operations in the system.

* Modify FiwareIdGenerator to be a utility class

FiwareIdGenerator has been updated to a final class with a private constructor that cannot be instantiated. This transformation ensures the class can only contain static methods, making it a utility class which is with respect to Java best practices.