Feature/refactoring to solve measurement issues (#31)
* Relocate model classes to 'cygnus' package

Model classes have been moved into a new sub-package 'cygnus' from the parent 'model' package for better organisation. Updated all import statements that were using the old package path in various files. This enhances the code structure and makes future development and maintenance easier.

* Remove Device entity and related services

Device entity, its related services and associated tests have been deleted. Also made updates in Validation logic in the Location and DeviceMeasurement classes. This is part of an ongoing refactor to simplify and restructure the codebase.

* Update project version in pom.xml

The version of the project within pom.xml has been updated from 6.4.0 to 7.0.0. This change is indicative of the ongoing major changes and refactorings being made to the codebase.

* Remove unused deviceMeasurementId in integration tests

The variable "deviceMeasurementId" that was previously generated and unused in both `DroneDeviceMeasurementIntegrationServiceIT` and `DeviceMeasurementIntegrationServiceIT` has been deleted. Tests are now using "deviceId" directly, simplifying the readability and efficiency of the code. Also, the warning suppression annotation was added to `StatusService` and `SubscriptionService` classes.