Feature/adapt fiware model to create valid entities (#34)
* Remove redundant FIWARE integration layer services

Deleted a number of redundant classes related to the FIWARE integration layer, including various services, models, enumeration, and test classes for these components. This deletion is part of a larger clean up effort to simplify the 5gLa FIWARE integration layer.

* Update version of 5gLa FIWARE integration layer

The version of the 5gLa FIWARE integration layer has been updated in the pom.xml file. The version has been increased from 8.0.0 to 9.0.0, as part of an overall software update.

* Remove unused imports and make key field final

Unused import statements have been removed from the AbstractIntegrationService class. In addition, the key field in the FiwareTypes class has been made final for immutable object design and improved thread safety.

* Mark deviceMeasurement field as final in DeviceMeasurementBuilder

The deviceMeasurement field in the DeviceMeasurementBuilder class has been marked as final to ensure the field cannot be modified. This change enhances the immutability of the class, contributing to improved thread safety. The unused import statements in the AbstractIntegrationService class were also cleaned up.

* Add metadata to attribute in DeviceMeasurementBuilder

The updated code adds a metadata list to the attribute in the DeviceMeasurementBuilder class. This is achieved by setting the metadata to the attribute of deviceMeasurement. The change enhances the flexibility of the attribute, allowing it to contain more information about the DeviceMeasurement.