Feature/add controlled property as metadata (#35)
* Update DeviceMeasurementBuilder to include metadata

An additional method was added to DeviceMeasurementBuilder to include metadata fields, adding flexibility for data inclusion in the attributes. A new "Text" key was also added to the FiwareTypes enumeration. These modifications will allow better handling of device measurements by carrying essential information about them.

* Remove metadata method from DeviceMeasurementBuilder

The metadata method that was responsible for setting the metadata fields in the DeviceMeasurementBuilder has been removed. This code change simplifies the DeviceMeasurementBuilder and removes potential redundancy as the metadata is currently not required in our use case.

* Remove metadata method from DeviceMeasurementBuilder

The metadata method that was responsible for setting the metadata fields in the DeviceMeasurementBuilder has been removed. This code change simplifies the DeviceMeasurementBuilder and removes potential redundancy as the metadata is currently not required in our use case.