Feature/allow multiple metadata entries (#36)
* Refactor DeviceMeasurementBuilder to use MetadataEntry instead of String attributes

In this update, the 'withMeasurement' method in the DeviceMeasurementBuilder class is refactored to accept 'MetadataEntry' objects rather than String attributes. The shift removes the static controlledProperty metadata attribute, replacing it with an optional, variable list of metadata entries, increasing modularity and flexibility in handling metadata.

* Update version in pom.xml

The version in the pom.xml file of the '5gLa FIWARE integration layer' project has been updated. The version number has been incremented from 9.1.0 to 9.2.0, reflecting the recent changes in the codebase.