Feature/fix npe in case there is no metadata (#38)
* Add null and empty check for metadata in Attribute.java

The method metadataAsJson() in Attribute.java has been updated to handle null and empty metadata by returning an empty string. Previously, if metadata was null or empty, an error could occur when attempting to parse it.

* Update version in pom.xml file

The version number of the application has been updated from 10.0.0 to 10.1.0 in the pom.xml file, indicating new adjustments or features have been implemented. This version change reflects updates to the application's functionality.

* The provided commit message does not match with the provided diff code. The appropriate commit message for the diff code would be:

Add metadata to DeviceMeasurementBuilder

Moved the instantiation of the metadata ArrayList in DeviceMeasurementBuilder. Now, it gets populated with "dateObservedMetadata" before handling additional metadata entries, ensuring "dateObservedMetadata" will always be part of the metadata list regardless of whether there are additional entries or not.

* Refactor variable name in DeviceMeasurementBuilder

Changed the name of the variable handling measurements from 'attribute' to 'measurement' in DeviceMeasurementBuilder class. This makes the usage clearer and more consistent with the logical function in the code context.