Replace Attribute type with new Location type for device measurements
The "location" in DeviceMeasurements has been updated from the generic Attribute type to use a new explicit Location class. Correspondingly, the DeviceMeasurementBuilder has been adjusted to work with the Location class when setting the location. The format for geographical information in FiwareTypes was updated from "geo:point" to "geo:json" to align with this change.