Feature/adapt location to create valid entries (#40)
* Replace Attribute type with new Location type for device measurements

The "location" in DeviceMeasurements has been updated from the generic Attribute type to use a new explicit Location class. Correspondingly, the DeviceMeasurementBuilder has been adjusted to work with the Location class when setting the location. The format for geographical information in FiwareTypes was updated from "geo:point" to "geo:json" to align with this change.

* Update version in pom.xml

The version of the "fiware-integration-layer" artifact in pom.xml has been updated from 10.2.0 to 10.3.0. This is a routine version bump as part of our product lifecycle management.