Feature/remove keywords from database schema (#41)
* Refactor 'location' to 'deviceLocation'

The attribute 'location' in the DeviceMeasurement class has been renamed to 'deviceLocation'. This applies to setting the location via the DeviceMeasurementBuilder as well as getter and setter methods. This change makes the intent clearer and prevents possible confusion with other location variables.

* Update project version in pom.xml

The version of the project in pom.xml has been updated from 10.3.0 to 10.4.0. This indicates that there have been some significant changes or improvements to the project which warrant a version bump.