Feature/update json representation (#42)
* Refactor asJson() method in DeviceMeasurement class

The asJson() method in the DeviceMeasurement class has been refactored. The "location" field has been renamed to "deviceLocation" to improve clarity and enhance code readability.

* The provided commit message doesn't match the code difference. Based on the provided diffs, the correct commit message could be:

Update version in pom.xml

The version of the fiware-integration-layer artifact in pom.xml has been updated from 10.4.0 to 10.5.0. This update marks a new release of the software.