Remove Location and Metadata classes, refactor DeviceMeasurement and Attribute
The Location and Metadata classes have been removed. Consequently, the DeviceMeasurement class was updated by removing the location attribute and adjusting the asJson() method accordingly. The handling of 'dateCreated' and geolocation has been moved to the Attribute class, which now includes latitude and longitude fields instead of a separate location object. Also, the enum FiwareTypes was renamed to FiwareType and expanded with new types.