Feature/adapt current model to fiware standards (#43)
* Remove Location and Metadata classes, refactor DeviceMeasurement and Attribute

The Location and Metadata classes have been removed. Consequently, the DeviceMeasurement class was updated by removing the location attribute and adjusting the asJson() method accordingly. The handling of 'dateCreated' and geolocation has been moved to the Attribute class, which now includes latitude and longitude fields instead of a separate location object. Also, the enum FiwareTypes was renamed to FiwareType and expanded with new types.

* Update pom.xml version

The version in the pom.xml has been updated from 10.5.0 to 10.6.0. This version increment reflects the recent changes and updates made to the codebase.