Feature/fix json format to create valid entities (#45)
* Refactor JSON output in Attribute model

The JSON output methods in the Attribute class have been refactored to improve readability and structure. The 'dateCreatedAsJson' and 'locationAsJson' methods now return their respective attributes enclosed in individual objects, enhancing the modularity and understanding of the output.

* Update version number in pom.xml

The version number in pom.xml file has been updated from 10.7.0 to 10.8.0, indicating a new release. This change marks the completion of new features, bug fixes, or improvements in the software, thus warranting a version increase.