Add type validation and refactor DeviceMeasurement building
Additional validation for the 'type' attribute was added to the validate method of the DeviceMeasurement class. Moreover, the building process of a DeviceMeasurement instance was refactored with the removal of the Attribute class dependency, directly assigning properties like 'name', 'value', 'dateCreated', latitude, longitude, and externalDataReference in the DeviceMeasurementBuilder. This refactoring simplifies the model's structure and its maintenance.