Update version in pom.xml
The version of the 'fiware-integration-layer' in the pom.xml file was updated from 10.8.0 to 11.0.0, preparing for a new release. This change reflects the recent modifications in the project, including the addition of type validation in the DeviceMeasurement class and the refactoring of its building process, as well as removing the dependency on the Attribute class.