Remove Attribute class from fiware model
The Attribute class from the fiware model has been deleted. This class was no longer necessary due to recent refactoring and changes in the structures, as its functionalities have been covered by other components. The project update also included other modifications, such as the version update in pom.xml, as part of the preparation for a new release.