Feature/adapt fiware model to create geospatial data (#46)
* Refactor DeviceMeasurement model class

The DeviceMeasurement class has been significantly modified by removing its dependency on the Attribute class. Several properties, including 'name', 'value', 'dateCreated', 'latitude', 'longitude' and 'externalDataReference', have been directly added and described in the DeviceMeasurement class. This update simplifies the structure of the model and makes it easier to interpret and maintain.

* Add type validation and refactor DeviceMeasurement building

Additional validation for the 'type' attribute was added to the validate method of the DeviceMeasurement class. Moreover, the building process of a DeviceMeasurement instance was refactored with the removal of the Attribute class dependency, directly assigning properties like 'name', 'value', 'dateCreated', latitude, longitude, and externalDataReference in the DeviceMeasurementBuilder. This refactoring simplifies the model's structure and its maintenance.

* Update version in pom.xml

The version of the 'fiware-integration-layer' in the pom.xml file was updated from 10.8.0 to 11.0.0, preparing for a new release. This change reflects the recent modifications in the project, including the addition of type validation in the DeviceMeasurement class and the refactoring of its building process, as well as removing the dependency on the Attribute class.

* Remove Attribute class from fiware model

The Attribute class from the fiware model has been deleted. This class was no longer necessary due to recent refactoring and changes in the structures, as its functionalities have been covered by other components. The project update also included other modifications, such as the version update in pom.xml, as part of the preparation for a new release.