Feature/check if values needs to be encoded (#47)
* Add jackson-core dependency and refactor JSON handling

The jackson-core dependency was added in `pom.xml` to enhance JSON handling in the application. The `DeviceMeasurement` class was refactored to handle null fields when generating JSON output. Helper methods were introduced for managing `externalDataReference`, `dateCreated`, and `location` fields. This change improves output cleanliness and standardizes data for users and systems.

* Update pom.xml version

The version of "fiware-integration-layer" has been updated in pom.xml from 11.0.0 to 11.1.0. This version bump is in line with the usual progress and software versioning convention.