Feature/use generic attribute to allow empty values (#50)
* Refactor DeviceMeasurement model attributes

Simplified variable types in the DeviceMeasurement model. Instead of having separate attribute types, all attributes are generalized under 'Attribute'. This makes the code cleaner and more maintainable, removing the need for specific attribute types.

* Update version in pom.xml

Increased the version of the FIWARE integration layer project in pom.xml from 12.0.0 to 12.1.0. This version bump corresponds to the