Feature/remove key value options from request (#51)
* Removal of redundant URI query parameter

This commit removes the "/op/update" + "?options=keyValues" query parameter from the HTTP request URI in the DeviceMeasurementIntegrationService. The "?options=keyValues" was found to be redundant and has been removed to clean-up the code and improve readability.

* Update version in pom.xml

This commit updates the version of the fiware-integration-layer artifact in the pom.xml from 12.1.0 to 12.2.0. This change reflects the ongoing evolution and progress of the project.