Feature/add device positions to data model (#53)
* Add DevicePosition model and integration service

This commit introduces a new DevicePosition model. It is used to store the position of a device within the FIWARE system. A DevicePositionIntegrationService has also been added for persisting the DevicePosition entities. Additionally, this commit refactors the DeviceMeasurement model to leverage the FiwareEntity interface for creating a JSON representation.

* Update type for 'entities' in UpdateOrCreateDeviceMeasurementRequest

The type of 'entities' in UpdateOrCreateDeviceMeasurementRequest has been changed from DeviceMeasurement to FiwareEntity. This change involved adjusting related methods to match the new type. It was also necessary to adjust imports as a consequence of this type change.

* Update type for 'entities' in UpdateOrCreateDeviceMeasurementRequest

The type of 'entities' in UpdateOrCreateDeviceMeasurementRequest has been changed from DeviceMeasurement to FiwareEntity. This change involved adjusting related methods to match the new type. It was also necessary to adjust imports as a consequence of this type change.

* Update project version in pom.xml

The project version in the pom.xml file has been updated from 13.0.0 to 14.0.0. This version increment indicates the completion of new features or significant changes to the existing codebase.