Bugfix/fix typo in service name (#54)
* Correct typo in class name

The class DevicePositiontIntegrationService has been renamed to DevicePositionIntegrationService to correct a typographical error. This change is purely cosmetic and does not affect the functionality of the class.

* Update version in pom.xml to 14.1.0

The version of the FIWARE integration layer in the pom.xml has been updated from 14.0.0 to 14.1.0, reflecting recent changes to the project. This update helps maintain accurate and up-to-date project versioning.